\name{MediaPackage-package}
\alias{MediaPackage-package}
\alias{MediaPackage}
\encoding{UTF-8}
\docType{package}
\title{MediaPackage}

\description{MediaPackage ha un unico obiettivo, quello di convogliare il calcolo dei principali tipi di medie in un unica libreria. La documentazione è chiara e scritta completamente in Italiano ed è possibile accedervi digitando in console il nome della funzione preceduto da "?". L'applicazione di queste funzioni è molto facile, ma allo stesso tempo utile in qualsiasi contesto, soprattutto in quello statistico-applicativo.}

\details{In questo Package è possibile trovare le seguenti funzioni: MediaAritmetica(), MediaArmonica(), MediaGeometrica() e MediaQuadratica().}

\note{Per qualsiasi informazione aggiuntiva, o malfunzionamento della libreria, si prega di contattare l'autore tramite email al seguente indirizzo: \email{antoniop.r003@gmail.com}}

\author{AntonioP-R}
