% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_create_channel}
\alias{mediapackage_create_channel}
\title{Creates a new Channel}
\usage{
mediapackage_create_channel(Description, Id, Tags)
}
\arguments{
\item{Description}{A short text description of the Channel.}

\item{Id}{[required] The ID of the Channel. The ID must be unique within the region and it
cannot be changed after a Channel is created.}

\item{Tags}{}
}
\description{
Creates a new Channel.
}
\section{Request syntax}{
\preformatted{mediapackage$create_channel(
  Description = "string",
  Id = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
